Extract DiscoveryBehaviour from Behaviour
