+++ core/network/src/lib.rs
@@ -27,6 +27,7 @@ mod behaviour;
 mod chain;
 mod custom_proto;
 mod debug_info;
+mod discovery;
 //mod libp2p_tests;
 mod on_demand_layer;
 #[macro_use]
